+++ packages/uicore/src/components/SelectWithSubmenu/SelectWithSubmenu.tsx
@@ -6,18 +6,14 @@
  */
 
 import React, { useCallback, useEffect, useState } from 'react'
-import cx from 'classnames'
-import { defaultTo, omit } from 'lodash-es'
-import { Menu, MenuItem, Popover, PopoverInteractionKind, Position } from '@blueprintjs/core'
+import { defaultTo } from 'lodash-es'
+import { Menu, MenuItem, Position } from '@blueprintjs/core'
 import { ISelectProps, Suggest } from '@blueprintjs/select'
 
 import { Text } from '../Text/Text'
 
 import css from './SelectWithSubmenu.css'
 import { Button } from '../../components/Button/Button'
-import selectCss from '../Select/Select.css'
-import { SelectProps, SelectOption, Select } from '../../components/Select/Select'
-import { MultiTypeInputType } from 'index'
 
 export interface SelectWithSubmenuOption {
   label: string
@@ -28,14 +24,9 @@ export interface SelectWithSubmenuOption {
   hasSubmenuItems?: boolean
 }
 
-export interface SubmenuSelectOption extends SelectOption {
-  submenuItems: SelectOption[]
-  hasSubItems?: boolean
-}
-
 type Props = ISelectProps<SelectWithSubmenuOption>
 
-export interface SelectWithSubmenuPropsV1
+export interface SelectWithSubmenuProps
   extends Omit<
     Props,
     | 'popoverProps'
@@ -53,21 +44,14 @@ export interface SelectWithSubmenuPropsV1
   onChange?(opts: SelectWithSubmenuOption): void
   onSubmenuOpen?: (item?: SelectWithSubmenuOption) => Promise<void>
 }
-export interface SelectWithSubmenuPropsV2 extends Omit<SelectProps, 'items' | 'onChange'> {
-  items: SubmenuSelectOption[]
-  loading?: boolean
-  onOpening?: (item: SelectOption) => void
-  interactionKind?: PopoverInteractionKind
-  onChange?: (primaryValue: SelectOption, selectedValue?: SelectOption, type?: MultiTypeInputType) => void
-}
 
-export function SelectWithSubmenuV1(props: SelectWithSubmenuPropsV1): React.ReactElement {
+export function SelectWithSubmenu(props: SelectWithSubmenuProps): React.ReactElement {
   const { items, onChange, onSubmenuOpen, allowCreatingNewItems, value, ...selectProps } = props
   const [selectedItem, setSelectedItem] = useState<SelectWithSubmenuOption | undefined>(undefined)
   const itemRenderer = useCallback(
     (item: SelectWithSubmenuOption, { handleClick }) => {
       return item.hasSubmenuItems ? (
-        <SubmenuV1
+        <Submenu
           items={defaultTo(item.submenuItems, [])}
           onChange={subItem => {
             onChange?.(subItem)
@@ -164,14 +148,14 @@ export function SelectWithSubmenuV1(props: SelectWithSubmenuPropsV1): React.Reac
   )
 }
 
-interface SubmenuPropsV1 {
+interface SubmenuProps {
   items: SelectWithSubmenuOption[]
   onChange?: (item: SelectWithSubmenuOption) => void
   primaryValue?: SelectWithSubmenuOption
   onSubmenuOpen?: (item?: SelectWithSubmenuOption) => Promise<void>
 }
 
-const SubmenuV1 = ({ items, onChange, primaryValue, onSubmenuOpen }: SubmenuPropsV1) => {
+const Submenu = ({ items, onChange, primaryValue, onSubmenuOpen }: SubmenuProps) => {
   const [loading, setLoading] = useState(false)
 
   useEffect(() => {
@@ -210,104 +194,5 @@ const SubmenuV1 = ({ items, onChange, primaryValue, onSubmenuOpen }: SubmenuProp
     </div>
   ) : (
     <div className={css.noResultsFound}>No Results Found</div>
-  )
-}
-
-interface SubmenuProps {
-  items: SelectOption[]
-  loading?: boolean
-  onChange?: (primaryValue: SelectOption, selectedValue: SelectOption, type?: MultiTypeInputType) => void
-  primaryValue?: SelectOption
-}
-const SubmenuV2 = ({ items, onChange, primaryValue, loading }: SubmenuProps) => {
-  return (
-    <Menu className={css.submenu}>
-      {loading ? (
-        <div className={css.noResultsFound}>Loading...</div>
-      ) : items?.length ? (
-        items.map((item: any) => (
-          <MenuItem
-            key={item.value}
-            text={item.label}
-            onClick={() => {
-              onChange?.(primaryValue as SelectOption, item)
-            }}
-            className={css.submenuItem}
-          />
-        ))
-      ) : (
-        <div className={css.noResultsFound}>No Results Found</div>
-      )}
-    </Menu>
-  )
-}
-
-export function SelectWithSubmenuV2(props: SelectWithSubmenuPropsV2) {
-  const { items, className, loading, onChange, ...selectProps } = props
-
-  const itemRenderer = useCallback(
-    (item: SelectOption) => {
-      return (item as SubmenuSelectOption).hasSubItems ? (
-        <Popover
-          content={
-            <SubmenuV2
-              items={(item as SubmenuSelectOption).submenuItems}
-              onChange={onChange}
-              primaryValue={omit(item, 'submenuItems')}
-              loading={loading}
-            />
-          }
-          position={Position.LEFT_TOP}
-          interactionKind={props?.interactionKind || PopoverInteractionKind.HOVER}
-          onOpening={() => {
-            if (props.onOpening) props.onOpening(item)
-          }}
-          minimal
-          usePortal
-          className={css.wrapperClassName}
-          targetClassName={css.targetClassName}
-          modifiers={{
-            preventOverflow: {
-              // This is required to always attach the popover portal to the start of the reference instead of the middle
-              escapeWithReference: true
-            },
-            offset: {
-              // This is required to offset the portal after it is attached to the reference.
-              // By default the portal is positioned at top: 0, left:0 wrt it's reference
-              // -2 is to adjust for the removed
-              offset: '0 2'
-            }
-          }}>
-          <li
-            key={item.value?.toString()}
-            className={cx(selectCss.menuItem, css.menuItem)}
-            onClick={event => event.preventDefault()}>
-            <Text rightIcon={'chevron-right'}>{item.label}</Text>
-          </li>
-        </Popover>
-      ) : (
-        <li
-          key={item.value?.toString()}
-          className={cx(selectCss.menuItem, css.menuItem)}
-          onClick={() => {
-            if (onChange) {
-              onChange(item)
-            }
-          }}>
-          <Text>{item.label}</Text>
-        </li>
-      )
-    },
-    [items, loading]
-  )
-
-  return (
-    <Select
-      {...selectProps}
-      items={items}
-      itemRenderer={itemRenderer}
-      className={className}
-      onChange={onChange as SelectProps['onChange']}
-    />
   )
 }
